pyproject.toml
[project]
name = "portfolio-media"
version = "0.1.0"
description = "AI-powered visual asset generation for GitHub packages"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CSF Development"}
]

dependencies = [
    "media-pipeline>=0.1.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/portfolio-media"
Repository = "https://github.com/portfolio-media"
Documentation = "https://github.com/portfolio-media#readme"
Issues = "https://github.com/portfolio-media/issues"

[project.optional-dependencies]
gemini = [
    "google-generativeai>=0.3.0",
]
openrouter = [
    "openai>=1.0.0",
]
glm = [
    "zhipuai>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "portfolio-media[gemini,openrouter,glm,dev]",
]

[project.scripts]
portfolio-media = "portfolio_media.cli:main"
portfolio-media-logo = "portfolio_media.logo_generator:generate_logo_cli"
portfolio-media-diagram = "portfolio_media.diagram_generator:generate_diagram_cli"
portfolio-media-screenshot = "portfolio_media.screenshot_capture:capture_screenshot_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/portfolio_media"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/skill",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
